Restructured module tree to be more standardised and to remove the './types' module
